# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a description of the Vihodnoy server.",
        "title": "Swagger Vihodnoy API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/login": {
            "post": {
                "description": "Авторизует пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Авторизация пользователя",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Wrong Credentials",
                        "schema": {
                            "$ref": "#/definitions/httpErrors.HttpError"
                        }
                    },
                    "401": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/utils.ValidationErrResponse"
                        }
                    },
                    "403": {
                        "description": "User is alredy logged in",
                        "schema": {
                            "$ref": "#/definitions/httpErrors.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpErrors.HttpError"
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "description": "Выход из аккаунта",
                "tags": [
                    "auth"
                ],
                "summary": "Выход из системы",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httpErrors.HttpError"
                        }
                    }
                }
            }
        },
        "/profile": {
            "get": {
                "description": "Возвращает информацию о профиле текущего пользователя",
                "tags": [
                    "profile"
                ],
                "summary": "Профиль пользователя",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.ProfileResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpErrors.HttpError"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "Создает нового пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Регистрация пользователя",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/auth.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid Data / Username or Email already taken",
                        "schema": {
                            "$ref": "#/definitions/httpErrors.HttpError"
                        }
                    },
                    "401": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/utils.ValidationErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpErrors.HttpError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "auth.ProfileResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "auth.UserResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "httpErrors.HttpError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.AuthError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "field": {
                    "type": "string"
                }
            }
        },
...
        "utils.ValidationErrResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AuthError"
                    }
                }
            }
        }
    }
}